implement data repair and reorganize modules
